requirements.txt
Django==5.2.4
psycopg2-binary==2.9.10
python-dotenv==1.0.0
djangorestframework==3.16.1